Completed Analysis

+++ Trady.Analysis/Indicator/ChandelierExit.cs
@@ -12,7 +12,7 @@ namespace Trady.Analysis.Indicator
         readonly LowestLowByTuple _ll;
         readonly AverageTrueRangeByTuple _atr;
 
-        public ChandelierExit(IEnumerable<TInput> inputs, Func<TInput, (decimal High, decimal Low, decimal Close)> inputMapper, Func<TInput, (decimal? Long, decimal? Short), TOutput> outputMapper, int periodCount, int atrCount) : base(inputs, inputMapper, outputMapper)
+        public ChandelierExit(IEnumerable<TInput> inputs, Func<TInput, (decimal High, decimal Low, decimal Close)> inputMapper, Func<TInput, (decimal? Long, decimal? Short), TOutput> outputMapper, int periodCount, decimal atrCount) : base(inputs, inputMapper, outputMapper)
         {
 			_hh = new HighestHighByTuple(inputs.Select(i => inputMapper(i).High), periodCount);
 			_ll = new LowestLowByTuple(inputs.Select(i => inputMapper(i).Low), periodCount);
@@ -37,7 +37,7 @@ namespace Trady.Analysis.Indicator
 
     public class ChandelierExitByTuple : ChandelierExit<(decimal High, decimal Low, decimal Close), (decimal? Long, decimal? Short)>
     {
-        public ChandelierExitByTuple(IEnumerable<(decimal High, decimal Low, decimal Close)> inputs, int periodCount, int atrCount) 
+        public ChandelierExitByTuple(IEnumerable<(decimal High, decimal Low, decimal Close)> inputs, int periodCount, decimal atrCount) 
             : base(inputs, i => i, (i, otm) => otm, periodCount, atrCount)
         {
         }
@@ -45,7 +45,7 @@ namespace Trady.Analysis.Indicator
 
     public class ChandelierExit : ChandelierExit<Candle, AnalyzableTick<(decimal? Long, decimal? Short)>>
     {
-        public ChandelierExit(IEnumerable<Candle> inputs, int periodCount, int atrCount) 
+        public ChandelierExit(IEnumerable<Candle> inputs, int periodCount, decimal atrCount) 
             : base(inputs, i => (i.High, i.Low, i.Close), (i, otm) => new AnalyzableTick<(decimal? Long, decimal? Short)>(i.DateTime, otm), periodCount, atrCount)
         {
         }
